api for the song app
1. serializers.py
add reverse relationships for models with their primary keys bieng used
as foreign keys in other tables

2. urls.py
add url specifications for the endpoints created in api/views.py

3. models.py
removed slug like urls in the Song class because it was bringing some
problems when trying to serialize the class. updated every field that
had null arg to also include blank to allow empty values to be sent in
the form.

4. permissions.py
add the permission to allow only owners of the lyrics, translation, and
lyric-requests to update and delete their posts

5. views.py
create endpoints for the song api

6. factories.py
remove slug field from the Song class so that it does not prompt django
to raise some exceptions

7. test_views.py
create test for new endpoints to make sure they work properly. the test
coverage was of 96%

8. project/urls.py
add the path to the song urls

9. config/settings/base.py
replace the old permission **IsAuthenticated** that was ristricting the user even from reading data if they are not logging with **IsAuthenticatedOrReadOnly** to overcome the unauthenticated reads. add DjangoFilterBackend to support data filtering.