- changes: swagger documentation link
1. backend/song/models.py
add a slug field to the Song class so that we can have a human readable
url. remove `__repr__(self)` to increase the coverage percentage. remove
prefixes on `related_name`

2. backend/song/api/serializers.py
change from PrimaryKeyRelatedField to HyperLinkedRelatedField to return
the url to the resource instead of the id this will reduce the overhead in the future

3. backend/song/api/urls.py
add the urlpattern for the swagger documentation. The docs are genarated
at the application level which means that each app will have its own
docs for its API

4. backend/song/tests/test_views.py
test for the existance of all the expected fields in response to a
request.

> drf_yasg is used to generate the swagger documentation for the song
app which can be accessed @`/api/v1/swagger` for more info checkout the
[drf_yasg docs](https://github.com/axnsan12/drf-yasg)